+++ src/routers/management/config.py
@@ -1,40 +1,72 @@
 from typing import Optional, Annotated, List
 from fastapi import APIRouter, Depends, Query
+from fastapi.exceptions import HTTPException
 from fastapi.responses import JSONResponse
-from src.models.env_config import EnvConfig, get_config
+from src.models.model_config import ModelConfig
+from src.models.chat_config import ChatConfig
 from src.controllers.providers import AIProvider
 from src.tasks.indexing import reset_index
+from src.controllers.env_configs import EnvConfigManager
 from create_llama.backend.app.settings import init_settings
 
 config_router = r = APIRouter()
 
 
-@r.get("")
-def get_current_config(
-    config: Annotated[EnvConfig, Depends(get_config)],
+@r.get("/is_configured")
+def is_configured(
+    config: ModelConfig = Depends(ModelConfig.get_config),
+) -> bool:
+    return config.configured
+
+
+@r.get("/chat", tags=["Chat config"])
+def get_chat_config(
+    config: ChatConfig = Depends(ChatConfig.get_config),
 ):
     return config.to_api_response()
 
 
-@r.post("")
-def update_config(
-    new_config: EnvConfig,
-    config: EnvConfig = Depends(get_config),
+@r.post("/chat", tags=["Chat config"])
+def update_chat_config(
+    new_config: ChatConfig,
+    config: ChatConfig = Depends(ChatConfig.get_config),
 ):
-    # Update config
-    new_config.to_runtime_env()
-    new_config.to_env_file()
+    EnvConfigManager.update(config, new_config, rollback_on_failure=True)
+
+    return JSONResponse(
+        {
+            "message": "Config updated successfully.",
+            "data": new_config.to_api_response(),
+        }
+    )
+
+
+@r.get("/models", tags=["Model config"])
+def get_model_config(
+    config: ModelConfig = Depends(ModelConfig.get_config),
+):
+    return config.to_api_response()
+
+
+@r.post("/models", tags=["Model config"])
+def update_model_config(
+    new_config: ModelConfig,
+    config: ModelConfig = Depends(ModelConfig.get_config),
+):
     # If the new config has a different model provider
-    # Or the AI config has not been configured yet
+    # Or the model config has not been configured yet
     # We need to:
     # 1. Reload the llama_index settings
     # 2. Reset the index
-    init_settings()
+    EnvConfigManager.update(config, new_config, rollback_on_failure=True)
+
+    # We won't rollback the changes if the index reset fails
     if (new_config.model_provider != config.model_provider) or not config.configured:
         reset_index()
 
     # Response with the updated config
-    config = get_config()
+    config = ModelConfig.get_config()
+
     return JSONResponse(
         {
             "message": "Config updated successfully.",
@@ -43,7 +75,7 @@ def update_config(
     )
 
 
-@r.get("/models")
+@r.get("/models/list", tags=["Model config"])
 def get_available_models(
     provider: Optional[str] = Query(
         None,